+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add a new formly type "section" to display a level 3 title (h3) inside a formly form
 - Add a new formly type "simple" to display a simple text next to a form label
+- Improve directive ov-tags and its corresponding formly type "tags". ov-tags now accepts a placeholder and a list of predefined values. If predefined values are set only those values can be added using the auto-complete
 
 ## BUG FIXES